+++ components/notification/index.tsx
@@ -1,11 +1,6 @@
 import * as React from 'react';
 import Notification from 'rc-notification';
-import {
-  NotificationInstance as RCNotificationInstance,
-  NoticeContent as RCNoticeContent,
-  HolderReadyCallback as RCHolderReadyCallback,
-} from 'rc-notification/lib/Notification';
-import useRCNotification from 'rc-notification/lib/useNotification';
+import { NotificationInstance as RCNotificationInstance } from 'rc-notification/lib/Notification';
 import {
   CloseOutlined,
   CheckCircleOutlined,
@@ -13,6 +8,7 @@ import {
   ExclamationCircleOutlined,
   InfoCircleOutlined,
 } from '@ant-design/icons';
+import createUseNotification from './hooks/useNotification';
 
 export type NotificationPlacement = 'topLeft' | 'topRight' | 'bottomLeft' | 'bottomRight';
 
@@ -239,40 +235,7 @@ const api: any = {
 });
 
 api.warn = api.warning;
-api.useNotification = (): [NotificationInstance, React.ReactElement] => {
-  // We create a proxy to handle delay created instance
-  let innerInstance: RCNotificationInstance | null = null;
-  const proxy = {
-    add: (noticeProps: RCNoticeContent, holderCallback?: RCHolderReadyCallback) => {
-      if (innerInstance) {
-        innerInstance.component.add(noticeProps, holderCallback);
-      }
-    },
-  } as any;
-
-  const [hookNotify, holder] = useRCNotification(proxy);
-
-  function notify(args: ArgsProps) {
-    getNotificationInstance(args).then(({ prefixCls, instance }) => {
-      innerInstance = instance;
-      hookNotify(getRCNoticeProps(args, prefixCls));
-    });
-  }
-
-  // Fill functions
-  const hookAPI: any = {
-    open: notify,
-  };
-  ['success', 'info', 'warning', 'error'].forEach(type => {
-    hookAPI[type] = (args: ArgsProps) =>
-      hookAPI.open({
-        ...args,
-        type,
-      });
-  });
-
-  return [hookAPI, holder];
-};
+api.useNotification = createUseNotification(getNotificationInstance, getRCNoticeProps);
 
 export interface NotificationInstance {
   success(args: ArgsProps): void;
